requirements.txt
fastapi
uvicorn
pymongo
PyJWT
pydantic-settings
langchain
langchain-openai
langchain-community
langchain-core